Safari-609.1.11.1/LayoutTests/fast/forms/state-save-of-detached-control-expected.txt

Test to NOT save state of a detached form control.

PASS document.getElementById("input2").value is ""
